+++ Open_Restaurante/Open_Restaurante/Open_Restaurante.csproj
@@ -73,10 +73,10 @@
     <Compile Include="Login\edit_Usuario.Designer.cs">
       <DependentUpon>edit_Usuario.cs</DependentUpon>
     </Compile>
-    <Compile Include="modificarProducto.cs">
+    <Compile Include="Productos\modificarProducto.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="modificarProducto.Designer.cs">
+    <Compile Include="Productos\modificarProducto.Designer.cs">
       <DependentUpon>modificarProducto.cs</DependentUpon>
     </Compile>
     <Compile Include="Productos\NewProducto.cs">
@@ -113,7 +113,7 @@
     <EmbeddedResource Include="Login\edit_Usuario.resx">
       <DependentUpon>edit_Usuario.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="modificarProducto.resx">
+    <EmbeddedResource Include="Productos\modificarProducto.resx">
       <DependentUpon>modificarProducto.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Productos\NewProducto.resx">
